test: Button 컴포넌트 스토리북 수정
