+++ frontend/src/components/WriterDocument.tsx
@@ -1,13 +1,9 @@
-import { Suspense } from 'react';
-
 import styled from 'styled-components';
 
 import useLevellogQuery from 'hooks/levellog/useLevellogQuery';
 import usePreQuestionQuery from 'hooks/preQuestion/usePreQuestionQuery';
 
-import Loading from 'pages/status/Loading';
-
-import Button from 'components/@commons/Button';
+import Button from 'components/@commons/button/Button';
 import UiViewer from 'components/@commons/markdownEditor/UiViewer';
 
 const WriterDocument = ({
