+++ frontend/src/components/header/ProfileDropdown.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Button from 'components/@commons/Button';
+import Button from 'components/@commons/button/Button';
 
 const ProfileDropdown = ({
   isShowProfileDropdown,
