+++ frontend/src/components/levellogs/LevellogViewModal.tsx
@@ -7,9 +7,9 @@ import useUser from 'hooks/useUser';
 import closeIcon from 'assets/images/close.svg';
 import { GITHUB_AVATAR_SIZE_LIST, TEAM_STATUS } from 'constants/constants';
 
-import Button from 'components/@commons/Button';
 import FlexBox from 'components/@commons/FlexBox';
 import Image from 'components/@commons/Image';
+import Button from 'components/@commons/button/Button';
 import UiViewer from 'components/@commons/markdownEditor/UiViewer';
 import ModalPortal from 'portal/ModalPortal';
 import { ParticipantType, TeamStatusType } from 'types/team';
