+++ frontend/src/components/preQuestion/PreQuestionViewModal.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components';
 
 import closeIcon from 'assets/images/close.svg';
 
-import Button from 'components/@commons/Button';
 import FlexBox from 'components/@commons/FlexBox';
 import Image from 'components/@commons/Image';
+import Button from 'components/@commons/button/Button';
 import UiViewer from 'components/@commons/markdownEditor/UiViewer';
 import ModalPortal from 'portal/ModalPortal';
 import { PreQuestionCustomHookType, PreQuestionFormatType } from 'types/preQuestion';
